fourchan: handle captcha query ticket parameter
A new query parameter, "ticket", is now required when requesting captcha. When a user requests captcha the first time, the server returns a json response with the ticket and the cooldown the user has to wait before the ticket is active. Then, the user has to add the ticket to the query parameters when requesting captcha.

This commit fixes "Invalid server response" error when getting captcha.