fourchan: fix sys uris
Change in e77b6b4 broke some sys uris (e.g. https://sys.4chan.org/v/post became https://sys.4chan.org/v/v/post). Before the change, the first parameter in `createSysUri` was used to determine whether a board is safe or not, but after the change, there is only `segments` parameter left, and because it is the vararg parameter, it treats the board name that was previously used for board safety check as a path segment. It leads to broken uri.